Remove singular general exception and define EIT
Although the assertion that Section 508 does not apply to national security systems was referenced, that is only one of several specified general exceptions under the law. While it may be beneficial to some readers to understand those exceptions this section currently describes electronic information technology (EIT) and should be referencing a definition so readers understand that includes software and applications, not just "services". 
Reference: § 1194.3   General exceptions via https://www.access-board.gov/guidelines-and-standards/communications-and-it/about-the-section-508-standards/section-508-standards for more information.